pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-receipt-audit"
version = "0.1.0"
description = "Multimodal receipt extraction and audit demo using Qwen3-VL + PEFT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "gradio",
  "torch",
  "torchvision",
  "transformers>=4.57.0",
  "peft",
  "accelerate",
  "bitsandbytes",
  "pillow",
  "safetensors",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
